+++ src/Pages/Asset/Components/ManagingAccounts.js
@@ -3,17 +3,22 @@ import { Table } from 'Components';
 import { useAssets } from 'redux/hooks';
 
 const ManagingAccounts = () => {
-  const { assetInfo: { managingAccounts: { managers = {} } = {} } = {}, assetInfoLoading: tableLoading } = useAssets();
+  const {
+    assetInfo: { managingAccounts: { allowGovControl, managers = {} } = {} } = {},
+    assetInfoLoading: tableLoading,
+  } = useAssets();
 
   const tableHeaders = [
     { displayName: 'Address', dataName: 'manager' },
     { displayName: 'Permissions', dataName: 'permissions' },
+    { displayName: 'Allow Governance', dataName: 'allowGovControl' },
   ];
 
   const tableData = Object.keys(managers || {}).reduce(
     (acc, curr) => [
       ...acc,
       {
+        allowGovControl,
         manager: curr,
         permissions: managers[curr],
       },
@@ -21,7 +26,14 @@ const ManagingAccounts = () => {
     []
   );
 
-  return <Table tableHeaders={tableHeaders} tableData={tableData} isLoading={tableLoading} title="Managing Accounts" />;
+  return (
+    <Table
+      tableHeaders={tableHeaders}
+      tableData={tableData}
+      isLoading={tableLoading}
+      title="Managing Accounts"
+    />
+  );
 };
 
 export default ManagingAccounts;